Extract rad remote helper into lib
