+++ git-helpers/src/lib.rs
@@ -20,3 +20,81 @@
 extern crate radicle_keystore as keystore;
 
 pub mod credential;
+
+use std::{
+    env,
+    io,
+    path::{Path, PathBuf},
+};
+
+use librad::{
+    git::local::{
+        transport::{LocalTransport, Localio, Mode::Stateful, Settings},
+        url::LocalUrl,
+    },
+    keys::{PublicKey, SecretKey},
+    paths::Paths,
+};
+use radicle_keystore::{crypto::Pwhash, FileStorage, Keystore};
+
+// FIXME: this should be defined elsewhere to be consistent between applications
+const SECRET_KEY_FILE: &str = "librad.key";
+
+pub fn run_rad_remote_helper() -> anyhow::Result<()> {
+    let url = {
+        let args = env::args().skip(1).take(2).collect::<Vec<_>>();
+        args[0]
+            .parse()
+            .or_else(|_| args[1].parse())
+            .or_else(|_| Err(anyhow::anyhow!("invalid args: {:?}", args)))
+    }?;
+
+    let git_dir = env::var("GIT_DIR").map(PathBuf::from)?;
+
+    let mut transport = {
+        let paths = Paths::from_env()?;
+        let key = get_signer(&git_dir, paths.keys_dir(), &url)?;
+        LocalTransport::new(Settings { paths, signer: key })
+    }?;
+
+    loop {
+        let mut buf = String::with_capacity(32);
+        io::stdin().read_line(&mut buf)?;
+        let line = buf.trim();
+
+        if line == "capabilities" {
+            println!("connect\n\n");
+            continue;
+        }
+
+        if let Some(service) = line.strip_prefix("connect ") {
+            let service = match service {
+                "git-upload-pack" => Ok(git2::transport::Service::UploadPack),
+                "git-receive-pack" => Ok(git2::transport::Service::ReceivePack),
+                unknown => Err(anyhow::anyhow!("unknown service: {}", unknown)),
+            }?;
+
+            println!();
+
+            transport
+                .connect(url, service, Stateful, Localio::inherit())?
+                .wait()?;
+
+            break;
+        }
+
+        return Err(anyhow::anyhow!("unexpected command: {}", line));
+    }
+
+    Ok(())
+}
+
+fn get_signer(git_dir: &Path, keys_dir: &Path, url: &LocalUrl) -> anyhow::Result<SecretKey> {
+    let pass = credential::Git::new(git_dir).get(url)?;
+    let file = keys_dir.join(SECRET_KEY_FILE);
+    let keystore = FileStorage::<_, PublicKey, _, _>::new(&file, Pwhash::new(pass));
+    keystore
+        .get_key()
+        .map(|keypair| keypair.secret_key)
+        .map_err(|e| e.into())
+}
